# [WIP] CUAEV aev constants memory optimization