--- a/torchani/cuaev/aev.cu
+++ b/torchani/cuaev/aev.cu
@@ -2,6 +2,7 @@
 #include <thrust/equal.h>
 #include <torch/extension.h>
 #include <cub/cub.cuh>
+#include <iostream>
 #include <vector>
 
 #include <ATen/Context.h>
@@ -26,6 +27,16 @@ constexpr int csubaev_offsets(int i, int j, int n) {
   return starting + offset;
 }
 
+inline
+cudaError_t checkCudaErrors(cudaError_t result)
+{
+  if (result != cudaSuccess) {
+    fprintf(stderr, "CUDA Runtime Error: %s\n", cudaGetErrorString(result));
+    assert(result == cudaSuccess);
+  }
+  return result;
+}
+
 #define MAX_NSPECIES 10
 #define RCR_X 0
 #define RCA_X 1
@@ -37,9 +48,20 @@ constexpr int csubaev_offsets(int i, int j, int n) {
 #define ETAR_X 7
 #define ETAA_X 8
 #define ZETA_X 9
-#define SUBAEV_OFFSET_X 10
+#define SHFR_N 10
+#define SHFA_N 11
+#define SHFZ_N 12
+#define SUBAEV_OFFSET_X 13
 
 __constant__ float AEV_CONSTANTS[SUBAEV_OFFSET_X + MAX_NSPECIES * MAX_NSPECIES];
+
+#define SHF_OFFSET 30  // Max num of shifts per ShfR_t, ShfA_t, ShfZ_t
+#define SHFR_X 0
+#define SHFA_X 1
+#define SHFZ_X 2
+#define N_SHF 3
+
+texture<float, 1, cudaReadModeElementType> AEV_PARAMETERS;
 
 void initAEVConsts(AEVScalarParams& aev_params, cudaStream_t stream) {
   int num_species = aev_params.num_species;
@@ -56,6 +78,9 @@ void initAEVConsts(AEVScalarParams& aev_params, cudaStream_t stream) {
   aev_constants[ETAR_X] = aev_params.EtaR_t[0].item<float>();
   aev_constants[ETAA_X] = aev_params.EtaA_t[0].item<float>();
   aev_constants[ZETA_X] = aev_params.Zeta_t[0].item<float>();
+  aev_constants[SHFR_N] = aev_params.ShfR_t.size(0);
+  aev_constants[SHFA_N] = aev_params.ShfA_t.size(0);
+  aev_constants[SHFZ_N] = aev_params.ShfZ_t.size(0);
   // precompute the aev offsets and load to constand memory
   for (int t = 0; t < num_species; ++t) {
     int offset = 0;
@@ -70,6 +95,36 @@ void initAEVConsts(AEVScalarParams& aev_params, cudaStream_t stream) {
   cudaMemcpyToSymbolAsync(
       AEV_CONSTANTS, aev_constants, sizeof(float) * aev_constants_size, 0, cudaMemcpyDefault, stream);
   delete[] aev_constants;
+
+  // texture parameters
+  cudaChannelFormatDesc channelDesc = cudaCreateChannelDesc(32, 0, 0, 0, cudaChannelFormatKindFloat);
+  AEV_PARAMETERS.addressMode[0] = cudaAddressModeClamp;
+  AEV_PARAMETERS.filterMode = cudaFilterModePoint;
+  AEV_PARAMETERS.normalized = false;
+
+  float* d_ShfR_t = reinterpret_cast<float*>(aev_params.ShfR_t.clone().detach().to(torch::kCPU).data_ptr());
+  float* d_ShfA_t = reinterpret_cast<float*>(aev_params.ShfA_t.clone().detach().to(torch::kCPU).data_ptr());
+  float* d_ShfZ_t = reinterpret_cast<float*>(aev_params.ShfZ_t.clone().detach().to(torch::kCPU).data_ptr());
+  float* h_aev_parameters = new float[N_SHF * SHF_OFFSET];
+  for (int t = 0; t < aev_params.ShfR_t.size(0); ++t) {
+    // h_aev_parameters[SHFR_X * SHF_OFFSET + t] = d_ShfR_t[t];
+    h_aev_parameters[SHFR_X * SHF_OFFSET + t] = aev_params.ShfR_t[t].item<float>();
+  }
+  for (int t = 0; t < aev_params.ShfA_t.size(0); ++t) {
+    h_aev_parameters[SHFA_X * SHF_OFFSET + t] = aev_params.ShfA_t[t].item<float>();
+  }
+  for (int t = 0; t < aev_params.ShfZ_t.size(0); ++t) {
+    h_aev_parameters[SHFZ_X * SHF_OFFSET + t] = aev_params.ShfZ_t[t].item<float>();
+  }
+  cudaArray* d_aev_parameters;
+  int mem_size = N_SHF * SHF_OFFSET * sizeof(float);
+  // checkCudaErrors(cudaMalloc(&d_aev_parameters, mem_size));
+  checkCudaErrors(cudaMallocArray(&d_aev_parameters, &channelDesc, N_SHF * SHF_OFFSET, 0));
+  // checkCudaErrors(cudaMemcpy(d_aev_parameters, h_aev_parameters, mem_size, cudaMemcpyHostToDevice));
+  checkCudaErrors(cudaMemcpyToArray(d_aev_parameters, 0, 0, h_aev_parameters, N_SHF * SHF_OFFSET * sizeof(float), cudaMemcpyHostToDevice));
+  checkCudaErrors(cudaBindTextureToArray(AEV_PARAMETERS, d_aev_parameters, channelDesc));
+  // size_t offset;
+  // cudaBindTexture2D(&offset, AEV_PARAMETERS, d_aev_parameters, channelDesc, width, height, pitch);
 }
 
 template <typename DataT>
@@ -248,10 +303,6 @@ template <typename SpeciesT, typename DataT, typename IndexT = int, int TILEX = 
 __global__ void cuAngularAEVs(
     torch::PackedTensorAccessor32<SpeciesT, 2, torch::RestrictPtrTraits> species_t,
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits> pos_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> ShfA_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> ShfZ_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> EtaA_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> Zeta_t,
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits> aev_t,
     PairDist<DataT>* d_Rij,
     PairDist<DataT>* d_centralAtom,
@@ -302,8 +353,8 @@ __global__ void cuAngularAEVs(
   DataT EtaA = AEV_CONSTANTS[ETAA_X];
   DataT Zeta = AEV_CONSTANTS[ZETA_X];
 
-  IndexT nShfA = ShfA_t.size(0);
-  IndexT nShfZ = ShfZ_t.size(0);
+  IndexT nShfA = AEV_CONSTANTS[SHFA_N];
+  IndexT nShfZ = AEV_CONSTANTS[SHFZ_N];
 
   PairDist<DataT> d = d_centralAtom[cIdx];
   int start_idx = d_centerAtomStartIdx[cIdx];
@@ -368,12 +419,14 @@ __global__ void cuAngularAEVs(
         IndexT subaev_offset = (int)AEV_CONSTANTS[SUBAEV_OFFSET_X + type_j * num_species + type_k];
 
         for (int itheta = tile.x; itheta < nShfZ; itheta += TILEX) {
-          DataT ShfZ = ShfZ_t[itheta];
+          // DataT ShfZ = ShfZ_t[itheta];
+          DataT ShfZ = tex1D(AEV_PARAMETERS, SHFZ_X * SHF_OFFSET + itheta);
 
           DataT factor1 = pow((1 + cos(theta_ijk - ShfZ)) / 2, Zeta);
 
           for (int ishfr = tile.y; ishfr < nShfA; ishfr += TILEY) {
-            DataT ShfA = ShfA_t[ishfr];
+            // DataT ShfA = ShfA_t[ishfr];
+            DataT ShfA = tex1D(AEV_PARAMETERS, SHFA_X * SHF_OFFSET + ishfr);
             DataT factor2 = exp(-EtaA * (Rijk - ShfA) * (Rijk - ShfA));
 
             DataT res = 2 * factor1 * factor2 * fc_ijk;
@@ -400,10 +453,6 @@ template <
 __global__ void cuAngularAEVs_backward_or_doublebackward(
     torch::PackedTensorAccessor32<SpeciesT, 2, torch::RestrictPtrTraits> species_t,
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits> pos_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> ShfA_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> ShfZ_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> EtaA_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> Zeta_t,
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits>
         grad_output, // for backward, this is daev, for double backward, this is dforce (i.e. ddcoord)
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits>
@@ -466,8 +515,8 @@ __global__ void cuAngularAEVs_backward_or_doublebackward(
   DataT EtaA = AEV_CONSTANTS[ETAA_X];
   DataT Zeta = AEV_CONSTANTS[ZETA_X];
 
-  IndexT nShfA = ShfA_t.size(0);
-  IndexT nShfZ = ShfZ_t.size(0);
+  IndexT nShfA = AEV_CONSTANTS[SHFA_N];
+  IndexT nShfZ = AEV_CONSTANTS[SHFZ_N];
 
   PairDist<DataT> d = d_centralAtom[cIdx];
   int start_idx = d_centerAtomStartIdx[cIdx];
@@ -572,14 +621,16 @@ __global__ void cuAngularAEVs_backward_or_doublebackward(
             (int)AEV_CONSTANTS[SUBAEV_OFFSET_X + type_j * (int)AEV_CONSTANTS[NUM_SPECIES_X] + type_k];
 
         for (int itheta = tile.x; itheta < nShfZ; itheta += TILEX) {
-          DataT ShfZ = ShfZ_t[itheta];
+          // DataT ShfZ = ShfZ_t[itheta];
+          DataT ShfZ = tex1D(AEV_PARAMETERS, SHFZ_X * SHF_OFFSET + itheta);
 
           DataT factor1 = pow((1 + cos(theta_ijk - ShfZ)) / 2, Zeta);
           DataT grad_factor1_theta = 1.0 / 2.0 * Zeta * pow((1 + cos(ShfZ - theta_ijk)) / 2, Zeta - 1) *
               sin(ShfZ - theta_ijk); // tricky 100ms improved
 
           for (int ishfr = tile.y; ishfr < nShfA; ishfr += TILEY) {
-            DataT ShfA = ShfA_t[ishfr];
+            // DataT ShfA = ShfA_t[ishfr];
+            DataT ShfA = tex1D(AEV_PARAMETERS, SHFA_X * SHF_OFFSET + ishfr);
             DataT factor2 = exp(-EtaA * (Rijk - ShfA) * (Rijk - ShfA));
             DataT grad_factor2_dist = -EtaA * (Rijk - ShfA) * factor2;
 
@@ -679,15 +730,13 @@ __global__ void cuAngularAEVs_backward_or_doublebackward(
 template <typename SpeciesT, typename DataT, int THREADS_PER_RIJ>
 __global__ void cuRadialAEVs(
     torch::PackedTensorAccessor32<SpeciesT, 2, torch::RestrictPtrTraits> species_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> ShfR_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> EtaR_t,
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits> aev_t,
     PairDist<DataT>* d_Rij,
     int nRadialRij) {
   int gidx = blockIdx.x * blockDim.x + threadIdx.x;
   int idx = gidx / THREADS_PER_RIJ;
 
-  int nShfR = ShfR_t.size(0);
+  int nShfR = AEV_CONSTANTS[SHFR_N];
   DataT EtaR = AEV_CONSTANTS[ETAR_X];
 
   if (idx >= nRadialRij)
@@ -709,8 +758,7 @@ __global__ void cuRadialAEVs(
   DataT fc = 0.5 * cos(PI * Rij / Rcr) + 0.5;
 
   for (int ishfr = laneIdx; ishfr < nShfR; ishfr += THREADS_PER_RIJ) {
-    DataT ShfR = ShfR_t[ishfr];
-
+    DataT ShfR = tex1D(AEV_PARAMETERS, SHFR_X * SHF_OFFSET + ishfr);
     DataT GmR = 0.25 * exp(-EtaR * (Rij - ShfR) * (Rij - ShfR)) * fc;
 
     atomicAdd(&aev_t[mol_idx][i][type_j * radial_sublength + ishfr], GmR);
@@ -721,8 +769,6 @@ __global__ void cuRadialAEVs(
 template <bool is_double_backward, typename SpeciesT, typename DataT, int THREADS_PER_RIJ>
 __global__ void cuRadialAEVs_backward_or_doublebackward(
     torch::PackedTensorAccessor32<SpeciesT, 2, torch::RestrictPtrTraits> species_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> ShfR_t,
-    torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits> EtaR_t,
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits>
         grad_aev, // daev for backward, ddaev for double backward
     torch::PackedTensorAccessor32<DataT, 1, torch::RestrictPtrTraits>
@@ -732,7 +778,7 @@ __global__ void cuRadialAEVs_backward_or_doublebackward(
   int gidx = blockIdx.x * blockDim.x + threadIdx.x;
   int idx = gidx / THREADS_PER_RIJ;
 
-  int nShfR = ShfR_t.size(0);
+  int nShfR = AEV_CONSTANTS[SHFR_N];
   DataT EtaR = AEV_CONSTANTS[ETAR_X];
 
   if (idx >= nRadialRij)
@@ -760,7 +806,7 @@ __global__ void cuRadialAEVs_backward_or_doublebackward(
   }
 
   for (int ishfr = laneIdx; ishfr < nShfR; ishfr += THREADS_PER_RIJ) {
-    DataT ShfR = ShfR_t[ishfr];
+    DataT ShfR = tex1D(AEV_PARAMETERS, SHFR_X * SHF_OFFSET + ishfr);
 
     DataT GmR = 0.25 * exp(-EtaR * (Rij - ShfR) * (Rij - ShfR));
     DataT GmR_grad = -EtaR * (-2 * ShfR + 2 * Rij) * GmR;
@@ -956,8 +1002,6 @@ Result cuaev_forward(const Tensor& coordinates_t, const Tensor& species_t, const
   int nblocks = (nRadialRij * 8 + block_size - 1) / block_size;
   cuRadialAEVs<int, float, 8><<<nblocks, block_size, 0, stream>>>(
       species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
-      aev_params.ShfR_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.EtaR_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       aev_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       d_radialRij,
       nRadialRij);
@@ -1012,10 +1056,6 @@ Result cuaev_forward(const Tensor& coordinates_t, const Tensor& species_t, const
     cuAngularAEVs<<<nblocks_angAEV, block_size, smem_size_aligned, stream>>>(
         species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
         coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
-        aev_params.ShfA_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-        aev_params.ShfZ_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-        aev_params.EtaA_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-        aev_params.Zeta_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
         aev_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
         d_angularRij,
         d_centralAtom,
@@ -1067,8 +1107,6 @@ Tensor cuaev_backward(const Tensor& grad_output, const AEVScalarParams& aev_para
   int nblocks = (result.nRadialRij * 8 + block_size - 1) / block_size;
   cuRadialAEVs_backward_or_doublebackward<false, int, float, 8><<<nblocks, block_size, 0, stream>>>(
       species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
-      aev_params.ShfR_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.EtaR_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       grad_output.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       grad_radial_dist.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       d_radialRij,
@@ -1103,10 +1141,6 @@ Tensor cuaev_backward(const Tensor& grad_output, const AEVScalarParams& aev_para
   cuAngularAEVs_backward_or_doublebackward<false><<<nblocks_angAEV, block_size, smem_size_aligned, stream>>>(
       species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
       coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
-      aev_params.ShfA_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.ShfZ_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.EtaA_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.Zeta_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       grad_output.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       grad_coord.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       d_angularRij,
@@ -1156,8 +1190,6 @@ Tensor cuaev_double_backward(const Tensor& grad_force, const AEVScalarParams& ae
   nblocks = (result.nRadialRij * 8 + block_size - 1) / block_size;
   cuRadialAEVs_backward_or_doublebackward<true, int, float, 8><<<nblocks, block_size, 0, stream>>>(
       species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
-      aev_params.ShfR_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.EtaR_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       grad_grad_aev.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       grad_force_coord_Rij.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       d_radialRij,
@@ -1182,10 +1214,6 @@ Tensor cuaev_double_backward(const Tensor& grad_force, const AEVScalarParams& ae
   cuAngularAEVs_backward_or_doublebackward<true><<<nblocks_angAEV, block_size, smem_size_aligned, stream>>>(
       species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
       coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
-      aev_params.ShfA_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.ShfZ_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.EtaA_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
-      aev_params.Zeta_t.packed_accessor32<float, 1, torch::RestrictPtrTraits>(),
       grad_force.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       grad_grad_aev.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
       d_angularRij,