--- a/torchani/cuaev/aev.h
+++ b/torchani/cuaev/aev.h
@@ -185,6 +185,7 @@ struct Result {
   }
 };
 
+void initAEVConsts(AEVScalarParams& aev_params, cudaStream_t stream);
 // cuda kernels
 Result cuaev_forward(const Tensor& coordinates_t, const Tensor& species_t, const AEVScalarParams& aev_params);
 Tensor cuaev_backward(const Tensor& grad_output, const AEVScalarParams& aev_params, const Result& result);
@@ -194,6 +195,7 @@ Tensor cuaev_double_backward(const Tensor& grad_force, const AEVScalarParams& ae
 // Only keep one copy of aev parameters
 struct CuaevComputer : torch::CustomClassHolder {
   AEVScalarParams aev_params;
+  bool aev_consts_initialized;
 
   CuaevComputer(
       double Rcr,
@@ -207,6 +209,11 @@ struct CuaevComputer : torch::CustomClassHolder {
       int64_t num_species);
 
   Result forward(const Tensor& coordinates_t, const Tensor& species_t) {
+    if (!aev_consts_initialized) {
+      cudaStream_t stream = at::cuda::getCurrentCUDAStream(coordinates_t.device().index());
+      initAEVConsts(aev_params, stream);
+      aev_consts_initialized = true;
+    }
     return cuaev_forward(coordinates_t, species_t, aev_params);
   }