--- a/torchani/cuaev/aev.h
+++ b/torchani/cuaev/aev.h
@@ -2,6 +2,7 @@
 
 #include <c10/cuda/CUDACachingAllocator.h>
 #include <torch/extension.h>
+#include <iostream>
 using torch::Tensor;
 using torch::autograd::AutogradContext;
 using torch::autograd::tensor_list;
@@ -186,6 +187,7 @@ struct Result {
 };
 
 void initAEVConsts(AEVScalarParams& aev_params, cudaStream_t stream);
+
 // cuda kernels
 Result cuaev_forward(const Tensor& coordinates_t, const Tensor& species_t, const AEVScalarParams& aev_params);
 Tensor cuaev_backward(const Tensor& grad_output, const AEVScalarParams& aev_params, const Result& result);
@@ -213,6 +215,7 @@ struct CuaevComputer : torch::CustomClassHolder {
       cudaStream_t stream = at::cuda::getCurrentCUDAStream(coordinates_t.device().index());
       initAEVConsts(aev_params, stream);
       aev_consts_initialized = true;
+      std::cout << "init constants\n";
     }
     return cuaev_forward(coordinates_t, species_t, aev_params);
   }