@yueyericardo 2021-03-05
After using share mem for ShfZ & ShfA
```
RUN                        Total AEV    Forward      Backward     Force        Optimizer    Others       Epoch time   GPU
0 cu Energy train          21.4 ms      12.7 ms      15.6 ms      0.0 ms       11.1 ms      3.4 ms       64.1 ms      1321.5MB
1 py Energy train          80.3 ms      12.5 ms      15.9 ms      0.0 ms       13.7 ms      3.6 ms       126.0 ms     2129.5MB
2 cu Energy + Force train  21.2 ms      12.6 ms      122.9 ms     72.1 ms      7.6 ms       5.0 ms       241.4 ms     2769.5MB
3 py Energy + Force train  79.8 ms      12.5 ms      433.0 ms     166.5 ms     8.0 ms       5.6 ms       705.4 ms     5673.5MB
```